at controllers.TicketManagementCotroller.java : 	added priceCalculation, 	requires to finish priceCalculation.setPriceInFlights(flights);
at entities.Flight.java :
	added price;

at services.PriceCalculation.java :
	deleted Lombok @Data annotation,
	renamed CONSTANT variables,
	changed Date data type to LocalDateTime and the procedure of days calculation,
	requires to finish setPriceInFlights(Set<Flight> flights);

at controllers.java
  changed System.out.println(e) on e.printStackTrace()

at webapp.ticketmanagement.jsp :
	added price object for the output;

Refactor data types in entities and methods:

		price 		Long
		lugagge 	Long
		luggagePrice	Long

		flightId 	Long
		planeId		Long
		airportId 	Long
		customerId 	Long
		ticketId	Long

    	airportSource		Long
   		airportDestination	Long

		dateTime 		LocalDateTime
		priority 				Boolean
		registrationPriority	Boolean

		planeLoadRate 	Double
		maxLoad 		Double
    	currentLoad 	Double

 Split mockito test class.

 Merge with Log4j logic.

+++ src/main/java/entities/Airport.java
@@ -8,6 +8,6 @@ import lombok.experimental.Accessors;
 @EqualsAndHashCode(exclude={"airportId"})
 @Accessors(chain = true)
 public class Airport {
-    private int airportId;
+    private Long airportId;
     private String name;
 }
